+++ lib/lightning/auth_providers/common.ex
@@ -22,7 +22,7 @@ defmodule Lightning.AuthProviders.Common do
       field :scope, :string
       field :instance_url, :string
       field :sandbox, :boolean, default: false
-      field :api_version, :string, default: nil
+      field :apiVersion, :string, default: nil
     end
 
     @doc """
@@ -54,7 +54,7 @@ defmodule Lightning.AuthProviders.Common do
         :scope,
         :instance_url,
         :sandbox,
-        :api_version
+        :apiVersion
       ])
       |> validate_required([:access_token, :refresh_token])
     end