+++ lib/lightning_web/live/credential_live/oauth_component.ex
@@ -5,6 +5,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
   import LightningWeb.OauthCredentialHelper
 
   alias Lightning.AuthProviders.Common
+  alias Lightning.AuthProviders.Common.TokenBody
   alias Lightning.Credentials
   alias Phoenix.LiveView.AsyncResult
 
@@ -20,9 +21,8 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
   attr :update_body, :any, required: true
   attr :action, :any, required: true
   attr :scopes_changed, :boolean, default: false
+  attr :sandbox_changed, :boolean, default: false
   attr :schema, :string, required: true
-  attr :sandbox_value, :boolean, default: false
-  attr :api_version, :string, default: ""
   slot :inner_block
 
   def fieldset(assigns) do
@@ -53,10 +53,9 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
            form: @form,
            action: @action,
            scopes_changed: @scopes_changed,
+           sandbox_changed: @sandbox_changed,
            token_body_changeset: @token_body_changeset,
            update_body: @update_body,
-           sandbox_value: @sandbox_value,
-           api_version: @api_version,
            schema: @schema,
            id: "inner-form-#{@id}"
          ],
@@ -90,10 +89,20 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
       display_authorize_button?(assigns, display_reauthorize_banner)
 
     display_userinfo =
-      display_userinfo?(assigns.oauth_progress, display_reauthorize_banner)
+      display_userinfo?(
+        assigns.oauth_progress,
+        display_reauthorize_banner,
+        assigns.scopes_changed,
+        assigns.sandbox_changed
+      )
 
     display_error =
-      display_error?(assigns.oauth_progress, display_reauthorize_banner)
+      display_error?(
+        assigns.oauth_progress,
+        display_reauthorize_banner,
+        assigns.scopes_changed,
+        assigns.sandbox_changed
+      )
 
     assigns =
       assigns
@@ -110,9 +119,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
         body_form <- Phoenix.HTML.FormData.to_form(:credential, @form, :body, [])
       }>
         <%= Phoenix.HTML.Form.hidden_input(body_form, :scope) %>
-        <%= Phoenix.HTML.Form.hidden_input(body_form, :sandbox) %>
         <%= Phoenix.HTML.Form.hidden_input(body_form, :expires_at) %>
-        <%= Phoenix.HTML.Form.hidden_input(body_form, :api_version) %>
         <%= Phoenix.HTML.Form.hidden_input(body_form, :access_token) %>
         <%= Phoenix.HTML.Form.hidden_input(body_form, :instance_url) %>
         <%= Phoenix.HTML.Form.hidden_input(body_form, :refresh_token) %>
@@ -565,8 +572,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
           id: _id,
           action: _action,
           scopes_changed: _scopes_changed,
-          sandbox_value: _sandbox_value,
-          api_version: _api_version,
+          sandbox_changed: _sandbox_changed,
           token_body_changeset: _changeset,
           update_body: _body,
           schema: _schema
@@ -602,18 +608,6 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
     handle_scopes_update(scopes, socket)
   end
 
-  def update(%{api_version: api_version}, socket) do
-    socket.assigns.token.result
-    |> token_to_params()
-    |> maybe_add_specific_provider_params(
-      socket.assigns,
-      socket.assigns.provider
-    )
-    |> socket.assigns.update_body.()
-
-    {:ok, socket |> assign(api_version: api_version)}
-  end
-
   def update(%{sandbox: sandbox}, socket) do
     wellknown_url = socket.assigns.adapter.wellknown_url(sandbox)
 
@@ -644,8 +638,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
            id: id,
            action: action,
            scopes_changed: scopes_changed,
-           sandbox_value: sandbox_value,
-           api_version: api_version,
+           sandbox_changed: sandbox_changed,
            token_body_changeset: token_body_changeset,
            update_body: update_body,
            schema: schema
@@ -653,6 +646,10 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
          token
        ) do
     adapter = Credentials.lookup_adapter(schema)
+
+    sandbox =
+      token_body_changeset
+      |> Ecto.Changeset.fetch_field!(:sandbox)
 
     socket
     |> assign(
@@ -662,11 +659,11 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
       token: AsyncResult.ok(token),
       update_body: update_body,
       adapter: adapter,
-      sandbox: sandbox_value,
-      api_version: api_version,
       provider: adapter.provider_name,
       action: action,
-      scopes_changed: scopes_changed
+      scopes_changed: scopes_changed,
+      sandbox_changed: sandbox_changed,
+      sandbox: sandbox
     )
   end
 
@@ -758,16 +755,12 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
       client: client,
       adapter: adapter,
       sandbox: sandbox,
-      provider: provider,
       update_body: update_body
     } = socket.assigns
 
     wellknown_url = adapter.wellknown_url(sandbox)
 
-    parsed_token =
-      token
-      |> token_to_params()
-      |> maybe_add_specific_provider_params(socket.assigns, provider)
+    parsed_token = token_to_params(token)
 
     update_body.(parsed_token)
 
@@ -787,6 +780,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :token_failed)
      |> assign(:scopes_changed, false)
+     |> assign(:sandbox_changed, false)
      |> assign(token: AsyncResult.failed(%AsyncResult{}, reason))}
   end
 
@@ -797,6 +791,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :refresh_failed)
      |> assign(:scopes_changed, false)
+     |> assign(:sandbox_changed, false)
      |> assign(token: AsyncResult.failed(%AsyncResult{}, reason))}
   end
 
@@ -807,6 +802,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :token_failed)
      |> assign(:scopes_changed, false)
+     |> assign(:sandbox_changed, false)
      |> assign(token: AsyncResult.failed(%AsyncResult{}, "Network error"))}
   end
 
@@ -815,6 +811,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :userinfo_received)
      |> assign(scopes_changed: false)
+     |> assign(:sandbox_changed, false)
      |> assign(userinfo: AsyncResult.ok(userinfo))}
   end
 
@@ -831,6 +828,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :userinfo_failed)
      |> assign(scopes_changed: false)
+     |> assign(:sandbox_changed, false)
      |> assign(userinfo: AsyncResult.failed(%AsyncResult{}, reason))}
   end
 
@@ -841,6 +839,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
      socket
      |> assign(:oauth_progress, :userinfo_failed)
      |> assign(:scopes_changed, false)
+     |> assign(:sandbox_changed, false)
      |> assign(userinfo: AsyncResult.failed(%AsyncResult{}, "Network error"))}
   end
 
@@ -905,7 +904,7 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
     end)
   end
 
-  defp params_to_token(%Lightning.AuthProviders.Common.TokenBody{} = token) do
+  defp params_to_token(%TokenBody{} = token) do
     struct!(
       OAuth2.AccessToken,
       token
@@ -916,16 +915,6 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
     )
   end
 
-  defp maybe_add_specific_provider_params(token_params, assigns, "Salesforce") do
-    assigns
-    |> Map.take([:sandbox, :api_version])
-    |> Map.merge(token_params)
-  end
-
-  defp maybe_add_specific_provider_params(token_params, _assigns, _provider) do
-    token_params
-  end
-
   defp display_loader?(oauth_progress) do
     oauth_progress not in List.flatten([:not_started | Map.values(@oauth_states)])
   end
@@ -933,16 +922,17 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
   defp display_reauthorize_banner?(%{
          action: action,
          scopes_changed: scopes_changed,
+         sandbox_changed: sandbox_changed,
          oauth_progress: oauth_progress
        }) do
     case action do
       :new ->
-        scopes_changed &&
+        (sandbox_changed || scopes_changed) &&
           oauth_progress in (@oauth_states.success ++
                                @oauth_states.failure)
 
       :edit ->
-        scopes_changed && oauth_progress not in [:started]
+        (sandbox_changed || scopes_changed) && oauth_progress not in [:started]
 
       _ ->
         false
@@ -957,12 +947,24 @@ defmodule LightningWeb.CredentialLive.OauthComponent do
       !display_reauthorize_banner
   end
 
-  defp display_userinfo?(oauth_progress, display_reauthorize_banner) do
-    oauth_progress == :userinfo_received && !display_reauthorize_banner
+  defp display_userinfo?(
+         oauth_progress,
+         display_reauthorize_banner,
+         scopes_changed,
+         sandbox_changed
+       ) do
+    oauth_progress == :userinfo_received && !display_reauthorize_banner &&
+      !(scopes_changed or sandbox_changed)
   end
 
-  defp display_error?(oauth_progress, display_reauthorize_banner) do
-    oauth_progress in @oauth_states.failure && !display_reauthorize_banner
+  defp display_error?(
+         oauth_progress,
+         display_reauthorize_banner,
+         scopes_changed,
+         sandbox_changed
+       ) do
+    oauth_progress in @oauth_states.failure && !display_reauthorize_banner &&
+      !(scopes_changed || sandbox_changed)
   end
 
   defp update_form(assigns) do