+++ CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to
 
 ### Fixed
 
+- Fixed issue where updating an existing Salesforce credential to use a
+  `sandbox` endpoint would not properly re-authenticate.
+  [#1842](https://github.com/OpenFn/lightning/issues/1842)
 - Navigate directly to settings from url hash and renders default panel
   when there is no hash.
   [#1971](https://github.com/OpenFn/lightning/issues/1971)